Moving DatabaseConfiguratin to :addon-api:database module
