+++ addon-api/database/src/com/thoughtworks/go/database/DatabaseConfiguration.java
@@ -1,5 +1,5 @@
-/*************************GO-LICENSE-START*********************************
- * Copyright 2014 ThoughtWorks, Inc.
+/*
+ * Copyright 2018 ThoughtWorks, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,21 +12,17 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *************************GO-LICENSE-END***********************************/
+ */
 
-package com.thoughtworks.go.configuration;
-
-import com.thoughtworks.go.util.SystemEnvironment;
+package com.thoughtworks.go.database;
 
 import java.util.Properties;
 
 public abstract class DatabaseConfiguration {
 
-    protected final SystemEnvironment systemEnvironment;
     protected final Properties properties;
 
-    protected DatabaseConfiguration(SystemEnvironment systemEnvironment) {
-        this.systemEnvironment = systemEnvironment;
+    protected DatabaseConfiguration() {
         this.properties = loadConfiguration();
     }
